+++ sample/application/controllers/tests/Dotenv.php
@@ -1,14 +1,9 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-use \X\Annotation\Access;
 use \X\Util\Logger;
 
-class DotenvTest extends AppController {
-
-  /**
-   * @Access(allow_login=true, allow_logoff=true)
-   */
+class Dotenv extends AppController {
   public function index() {
     Logger::print($_ENV);
   }