+++ sample/application/controllers/tests/FileHelper.php
@@ -1,15 +1,10 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-use \X\Annotation\Access;
 use \X\Util\FileHelper;
 use \X\Util\Logger;
 
-class FileHelperTest extends AppController {
-
-  /**
-   * @Access(allow_login=true, allow_logoff=true)
-   */
+class FileHelper extends AppController {
   public function index() {
     try {
       chdir(APPPATH . 'test_data');