+++ sample/application/controllers/tests/QueryCache.php
@@ -1,16 +1,10 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-use \X\Annotation\Access;
 use \X\Util\Logger;
 
-class QueryCacheTest extends AppController {
-
+class QueryCache extends AppController {
   protected $model = 'UserModel';
-
-  /**
-   * @Access(allow_login=true, allow_logoff=true)
-   */
   public function index() {
     try {
       // Cache the results of this search query.
@@ -35,9 +29,6 @@ class QueryCacheTest extends AppController {
     }
   }
 
-  /**
-   * @Access(allow_login=true, allow_logoff=true)
-   */
   public function deleteCache() {
     $this->UserModel->cache_delete('QueryCacheTest', 'index');
   }